+++ packages/segment-session-replay-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.43](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.42...@amplitude/segment-session-replay-plugin@0.0.43) (2026-08-31)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
+## [0.0.42](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.41...@amplitude/segment-session-replay-plugin@0.0.42) (2026-08-27)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
 ## [0.0.41](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.40...@amplitude/segment-session-replay-plugin@0.0.41) (2026-08-24)
 
 **Note:** Version bump only for package @amplitude/segment-session-replay-plugin